db-api/src/requirements.txt
flask
redis
google-cloud-bigtable==2.17.0
google-cloud-core==2.3.2